+++ index.js
@@ -8,8 +8,8 @@ Machine = exports.Machine = require('./lib/machine').Machine;
 
 exports.Item = require('./lib/item').Item;
 
-exports.StateStore = require('./lib/statestore/base').StateStore;
-exports.JsonStore = require('./lib/statestore/json').JsonStore;
+exports.CheckpointStore = require('./lib/storage/base').CheckpointStore;
+exports.JsonStore = require('./lib/storage/json').JsonStore;
 
 
 exports.jobmanager = require('./lib/jobmanager');